Edit item form implemented.
Now we can also edit a certain item given a name and a type. We still
do not get track of the order, but we will in a close future. Also, we
had problems using the WP function for updating a table, so we deleted
the entries and then re-insert them. I know and I feel bad for it, but
so far it is working at least. I'll fix it later. Also, all the other
problems.