Refactoring, language support, new icon and desktop enhancements.
Now we can have the plugin also in Spanish and other languages. 

Also, we reorganized our files in several folders (img, languages,
simple_html_dom...) and we changed our desktop and plugin icon. We have
a bug still for fixing: "Official website" is not translated when
inserting into our database.